Retry do_orders if menus are unavailable
If the api is unable to retrieve the menus from Waiter, or if
the menus are the same as the last time they were retrieved
(ie they have not been updated yet) do_orders delays and retries,
up to a maximum number of tries. If still unsuccessful it then
raises an exception.