Get correct cart ID from webpage
The cart IDs are ordered by day, and by menu within day. Our hack
of picking the first ID in the sorted list wasn't working.
Instead, we sort the menu IDs for a given day, find the index
of the desired menu from which to order, and return the
corresponding cart ID